Address PR issues (more refactoring)
