+++ packages/web-app-files/src/helpers/resource/actions/types.ts
@@ -1,0 +1,4 @@
+export enum TransferType {
+  COPY,
+  MOVE
+}
