+++ packages/web-app-files/src/helpers/resource/conflictHandling/types.ts
@@ -1,0 +1,10 @@
+export enum ResolveStrategy {
+  SKIP,
+  REPLACE,
+  KEEP_BOTH,
+  MERGE
+}
+export interface ResolveConflict {
+  strategy: ResolveStrategy
+  doForAllConflicts: boolean
+}
