+++ packages/web-app-files/src/helpers/resource/index.ts
@@ -1,6 +1,5 @@
+export * from './actions'
 export * from './filter'
 export * from './resource'
 export * from './sameResource'
-export * from './copyMove'
-export * from './conflict-handling/conflictDialog'
-export * from './conflict-handling/conflictUtils'
+export * from './conflictHandling'
